test(pytest) for crud fuctionality on video services

+++ app/videos/routers.py
@@ -1,6 +1,6 @@
 from typing import Optional
 import uuid
-from fastapi import APIRouter, HTTPException, Request, Form, Depends
+from fastapi import APIRouter, HTTPException, Request, Form, Depends, status
 from fastapi.responses import HTMLResponse
 
 from app import utils
@@ -87,6 +87,10 @@ def video_detail_view(request: Request, host_id:str):
 @login_required
 def video_edit_view(request: Request, host_id: str):
     obj = get_object_or_404(Video, host_id=host_id)
+    
+    if obj.user_id != request.user.username:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
+    
     context = {
         "object": obj
     }
@@ -104,6 +108,10 @@ def video_edit_post_view(request: Request, host_id: str, is_htmx=Depends(is_htmx
         "user_id": request.user.username
     }
     obj = get_object_or_404(Video, host_id=host_id)
+    
+    if obj.user_id != request.user.username:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
+    
     data, errors = utils.valid_schema_data_or_error(raw_data, VideoEditSchema)
     if len(errors) > 0:
         return render(request, "videos/edit.html", context, status_code=400)
@@ -128,7 +136,9 @@ def video_hx_edit_view(request: Request, host_id: str, is_htmx=Depends(is_htmx))
     except:
         not_found = True
     if not_found:
-        return HTMLResponse("Not found, please try again.")
+        return HTMLResponse("Not found, please try again.", status_code=404)
+    if obj.user_id != uuid.UUID(request.user.username):
+        return HTMLResponse("Not authorized to perform requested action", status_code=403)
     context = {
         "object": obj
     }
@@ -151,6 +161,8 @@ def video_hx_edit_post_view(request: Request, host_id: str, is_htmx=Depends(is_h
         not_found = True
     if not_found:
         return HTMLResponse("Not found, please try again.", status_code=404)
+    if obj.user_id != uuid.UUID(request.user.username):
+        return HTMLResponse("Not authorized to perform requested action", status_code=403)
     if delete:
         obj.delete()
         return HTMLResponse('Item Deleted', status_code=204)
